+++ grunt/config/if.js
@@ -41,9 +41,9 @@ module.exports = function (grunt) {
 				'string-replace:changelogTag',  // Update the changelog [add tag]
 				'string-replace:jekyllVersion', // Update the config [add tag]
 				'angular.release',              // Create the Angular release
-				'preprocess:manifest',          // Create a new manifest into docs folder
 				'copy:jekyll',                  // Create a new jekyll config into docs folder
 				'copy:index',                   // Create the layout file for jekyll
+				'preprocess:manifest',          // Create a new manifest into docs folder
 				'internal.bump',                // Update the package [add tag] and commit all files
 				'gitpush:develop',              // Push the commit
 				'gitcheckout:master',           // Go to the master branch
@@ -65,9 +65,9 @@ module.exports = function (grunt) {
 				'string-replace:changelogTag',
 				'string-replace:jekyllVersion',
 				'angular.release',
-				'preprocess:manifest',
 				'copy:jekyll',
-				'copy:index'
+				'copy:index',
+				'preprocess:manifest'
 			]
 		}
 	};